backend/requirements.txt
Django==5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
